% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspols_plot_metros.R
\name{uspols_plot_metros}
\alias{uspols_plot_metros}
\title{Query CELEX database}
\usage{
uspols_plot_metros(sub_geos, geo)
}
\description{
These functions make life easier when querying the behavioral datasets.
}
